requirements.txt
langchain==0.1.17
langchain-core==0.1.48
langchain-groq==0.1.0
streamlit==1.32.0
python-dotenv==1.0.1
requests==2.31.0
pandas==2.1.3
pyarrow==14.0.1
pydantic==2.5.3
